Backend/requirements.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
stripe==7.10.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
